% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials-package.R
\docType{package}
\name{trials-package}
\alias{trials}
\alias{trials-package}
\title{trials: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MikeHedde/trials}
  \item Report bugs at \url{https://github.com/MikeHedde/trials/issues}
}

}
\author{
\strong{Maintainer}: Mike Hedde \email{mickael.hedde@inrae.fr} (\href{https://orcid.org/0000-0002-6733-3622}{ORCID}) [copyright holder]

}
\keyword{internal}
